backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-multipart
pymupdf
python-dotenv
pydantic
yfinance
httpx
